+++ frontend/src/types/generalTypes.ts
@@ -1,3 +1,7 @@
+"use client";
+
+import { FrontendErrorDetails } from "src/types/apiResponseTypes";
+
 export interface LayoutProps {
   children: React.ReactNode;
   params: Promise<{
@@ -7,4 +11,12 @@ export interface LayoutProps {
 
 export interface OptionalStringDict {
   [key: string]: string | undefined;
+}
+
+export interface ParsedError {
+  message?: string;
+  searchInputs?: OptionalStringDict;
+  status?: number;
+  type?: string;
+  details?: FrontendErrorDetails;
 }
